Encrypt TLS between ELB and Fargate instances
From AWS documentation:
`The load balancer establishes TLS connections with the targets using
certificates that you install on the targets. The load balancer does
not validate these certificates. Therefore, you can use self-signed
certificates or certificates that have expired. Because the load
balancer is in a virtual private cloud (VPC), traffic between the load
balancer and the targets is authenticated at the packet level, so it is
not at risk of man-in-the-middle attacks or spoofing even if the
certificates on the targets are not valid.`

There are other attack vectors but they can be made very difficult to
achieve via policy.