Resolve Image tags to SHA
When deploying a tag is often supplied, but that makes it hard to see
the exact version running in ECS, as the container definition only
shows the tag (eg latest) not the SHA it resolves to. Also if :latest
is used for the image then cloudformation will refuse to update the
stack if no other changes are made, as it will detect an empty change
set.