Catch admin password admin secret pending deletion
The user can supply an existing secret to the prepare script which
will check that secret exists before proceeding. However if the secret
is pending deletion, the exists check will succeed but the deployment
will fail as the secret is unreadable.

This commit ensures the prepare script fails if a secret pending
deletion is provided.