Remove autoamted backups on delete
This is to relieve db quota issues in the ci account, as short lived clusters
were retaining backups for 14 days. These backups count as RDS instances for
quota, so caused the CI account to hit it's RDS instance quota.

Deleting backups on termintation is not a problem for disaster recovery
as the final snapshot is retained.