Set low ALB timeout
Its important that the ALB (Application Load Balancer) idle timeout is
lower than the target's timeout (puma). Otherwise the webserver may
timeout a session at the same time the lb is sending a request, which
leads to seemingly random 502 errors. Puma has a default timeout of 20s,
so 10s should be safe for the ALB.

Related: conjurinc/ops#778