Don't re-encode Data Key
The container startup script base64 encodes the data key, which works
fine when the data key is a random string generated by asm. If the
data key is already base64 encoded (eg from another instance) then
it gets base64 encoded again.

This commit prevents the double encoding by checking the length. It also
ensures that containers exit early if the data key is not 32 bytes.

Previously if an invalid data key was supplied, the containers would
stay running but admin login attempts would fail as the account 'conjur'
would fail to create. Having the containers exit early sends a strong
signal to developers that something is up, so they can read the logs and
find the message about data key length.