Increase RDS max_connections
After sustained running, the connection limit was hit leading to errors
from non-ecs clients attempting to get DB connections. This was due
to a misreading of the RDS docs that lead me to think max_connections
was always atleast 5000. The current max connections for a 4gb instance
is ~350. This commit makes the value configurable but sets the default
at 2000 which fits with the other defaults.

Related: conjurinc/ops#790